+++ tests/framework/web/ResponseTest.php
@@ -55,11 +55,10 @@ class ResponseTest extends \yiiunit\TestCase
 
         $this->assertEquals($expectedContent, $content);
         $this->assertEquals(206, $this->response->statusCode);
-        $headers = $this->response->headers;
-        $this->assertEquals('bytes', $headers->get('Accept-Ranges'));
-        $this->assertEquals('bytes ' . $expectedHeader . '/' . StringHelper::byteLength($fullContent), $headers->get('Content-Range'));
-        $this->assertEquals('text/plain', $headers->get('Content-Type'));
-        $this->assertEquals("$length", $headers->get('Content-Length'));
+        $this->assertEquals(['bytes'], $this->response->getHeader('Accept-Ranges'));
+        $this->assertEquals(['bytes ' . $expectedHeader . '/' . StringHelper::byteLength($fullContent)], $this->response->getHeader('Content-Range'));
+        $this->assertEquals(['text/plain'], $this->response->getHeader('Content-Type'));
+        $this->assertEquals(["$length"], $this->response->getHeader('Content-Length'));
     }
 
     public function wrongRanges()
@@ -104,27 +103,26 @@ class ResponseTest extends \yiiunit\TestCase
 
         static::assertEquals('test', $content);
         static::assertEquals(200, $this->response->statusCode);
-        $headers = $this->response->headers;
-        static::assertEquals('application/octet-stream', $headers->get('Content-Type'));
-        static::assertEquals('attachment; filename="test.txt"', $headers->get('Content-Disposition'));
-        static::assertEquals(4, $headers->get('Content-Length'));
+        static::assertEquals(['application/octet-stream'], $this->response->getHeader('Content-Type'));
+        static::assertEquals(['attachment; filename="test.txt"'], $this->response->getHeader('Content-Disposition'));
+        static::assertEquals([4], $this->response->getHeader('Content-Length'));
     }
 
     public function testRedirect()
     {
         $_SERVER['REQUEST_URI'] = 'http://test-domain.com/';
-        $this->assertEquals($this->response->redirect('')->headers->get('location'), '/');
-        $this->assertEquals($this->response->redirect('http://some-external-domain.com')->headers->get('location'), 'http://some-external-domain.com');
-        $this->assertEquals($this->response->redirect('/')->headers->get('location'), '/');
-        $this->assertEquals($this->response->redirect('/something-relative')->headers->get('location'), '/something-relative');
-        $this->assertEquals($this->response->redirect(['/'])->headers->get('location'), '/index.php?r=');
-        $this->assertEquals($this->response->redirect(['view'])->headers->get('location'), '/index.php?r=view');
-        $this->assertEquals($this->response->redirect(['/controller'])->headers->get('location'), '/index.php?r=controller');
-        $this->assertEquals($this->response->redirect(['/controller/index'])->headers->get('location'), '/index.php?r=controller%2Findex');
-        $this->assertEquals($this->response->redirect(['//controller/index'])->headers->get('location'), '/index.php?r=controller%2Findex');
-        $this->assertEquals($this->response->redirect(['//controller/index', 'id' => 3])->headers->get('location'), '/index.php?r=controller%2Findex&id=3');
-        $this->assertEquals($this->response->redirect(['//controller/index', 'id_1' => 3, 'id_2' => 4])->headers->get('location'), '/index.php?r=controller%2Findex&id_1=3&id_2=4');
-        $this->assertEquals($this->response->redirect(['//controller/index', 'slug' => 'äöüß!"§$%&/()'])->headers->get('location'), '/index.php?r=controller%2Findex&slug=%C3%A4%C3%B6%C3%BC%C3%9F%21%22%C2%A7%24%25%26%2F%28%29');
+        $this->assertEquals($this->response->redirect('')->getHeader('location'), ['/']);
+        $this->assertEquals($this->response->redirect('http://some-external-domain.com')->getHeader('location'), ['http://some-external-domain.com']);
+        $this->assertEquals($this->response->redirect('/')->getHeader('location'), ['/']);
+        $this->assertEquals($this->response->redirect('/something-relative')->getHeader('location'), ['/something-relative']);
+        $this->assertEquals($this->response->redirect(['/'])->getHeader('location'), ['/index.php?r=']);
+        $this->assertEquals($this->response->redirect(['view'])->getHeader('location'), ['/index.php?r=view']);
+        $this->assertEquals($this->response->redirect(['/controller'])->getHeader('location'), ['/index.php?r=controller']);
+        $this->assertEquals($this->response->redirect(['/controller/index'])->getHeader('location'), ['/index.php?r=controller%2Findex']);
+        $this->assertEquals($this->response->redirect(['//controller/index'])->getHeader('location'), ['/index.php?r=controller%2Findex']);
+        $this->assertEquals($this->response->redirect(['//controller/index', 'id' => 3])->getHeader('location'), ['/index.php?r=controller%2Findex&id=3']);
+        $this->assertEquals($this->response->redirect(['//controller/index', 'id_1' => 3, 'id_2' => 4])->getHeader('location'), ['/index.php?r=controller%2Findex&id_1=3&id_2=4']);
+        $this->assertEquals($this->response->redirect(['//controller/index', 'slug' => 'äöüß!"§$%&/()'])->getHeader('location'), ['/index.php?r=controller%2Findex&slug=%C3%A4%C3%B6%C3%BC%C3%9F%21%22%C2%A7%24%25%26%2F%28%29']);
     }
 
     /**